rename .scss module files to .mod.scss, fix global css problems, fix styling for new elements on existing pages
