+++ src/components/Form/OutcomeSelection/OutcomeBar.mod.scss
@@ -16,6 +16,7 @@
   .wrapper {
     position: relative;
     width: 100%;
+    display: flex;
   }
 
   input[type="radio"] {
