+++ src/components/Form/TextInputAdornment.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import className from 'classnames/bind'
 
-import styles from './TextInputAdornment.scss'
+import styles from './TextInputAdornment.mod.scss'
 
 const cx = className.bind(styles)
